feat: update bcs mod
