+++ sdk/typescript/package.json
@@ -16,8 +16,8 @@
     "doc": "typedoc",
     "vitest": "vitest",
     "test": "pnpm test:unit",
-    "test:unit": "vitest",
-    "test:e2e": "wait-on http://127.0.0.1:9123 -l --timeout 120000 && vitest run e2e",
+    "test:unit": "vitest src",
+    "test:e2e": "wait-on http://127.0.0.1:9123 -l --timeout 120000 && vitest test/e2e",
     "test:e2e:nowait": "vitest run e2e",
     "prepare:e2e": "cargo run --bin rooch",
     "prepublishOnly": "pnpm build",
