+++ sdk/typescript/src/bcs/index.ts
@@ -1,7 +1,6 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-export * from "./types"
-export * from "./serializer"
-export * from "./deserializer"
-export * from "./helper"
+export * from "../generated/runtime/bcs/mod";
+export * from "../generated/runtime/serde/mod";
+export * as roochBcsTypes from "../generated/runtime/rooch_types/mod";
