--- sdk/typescript/src/bcs/types.ts
@@ -1,13 +1,0 @@
-// Copyright (c) RoochNetwork
-// SPDX-License-Identifier: Apache-2.0
-
-export type Seq<T> = T[]
-
-export type Uint8 = number
-export type Uint16 = number
-export type Uint32 = number
-export type Uint64 = bigint
-export type Uint128 = bigint
-export type Uint256 = bigint
-export type AnyNumber = bigint | number
-export type Bytes = Uint8Array
