+++ sdk/typescript/src/index.ts
@@ -1,9 +1,8 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-export * from "./account"
-export * from "./bcs"
-export * from "./provider"
-export * from "./utils";
 export * from "./types";
-export * as Types from "./generated/client/types";
+export * from "./utils";
+export * from "./bcs";
+export * from "./account";
+export * from "./provider";
