+++ sdk/typescript/src/utils/index.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from "./key"
-export * from "../types/hexString"
+export * from "../types/hex"
 export * from "./memoize-decorator"
 export * from "./mixin"
 export * from "./hex";
