+++ sdk/typescript/tsconfig.json
@@ -17,6 +17,6 @@
     "target": "es2020",
     "pretty": true
   },
-  "include": ["src"],
+  "include": ["src", "test"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
